Add appt documentation, fix date parsing for appts
- Add basic documentation for appt to Developer Guide
- Move UpcomingAppointment#isValidDate() to Appointment so that it
  can be used for parsing PastAppointments too.
- Check for the presence of date and/or diagnosis tags when past
  and/or upcoming appointments are generated.
