+++ src/main/java/seedu/address/model/appointment/UpcomingAppointment.java
@@ -2,13 +2,11 @@ package seedu.address.model.appointment;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 /**
  * Represents an upcoming appointment for a patient.
  */
 public class UpcomingAppointment extends Appointment {
-    public static final String MESSAGE_CONSTRAINTS = "Date string should be of the format dd-MM-yyyy.";
     public final String value;
     /**
      * Constructs an {@code UpcomingAppointment} for a {@code Patient}.
@@ -31,20 +29,6 @@ public class UpcomingAppointment extends Appointment {
         this.value = dateString;
     }
 
-    /**
-     * Returns true if a given string is a valid date.
-     */
-    public static boolean isValidDate(String test) {
-        if (test == null) {
-            return true;
-        }
-        try {
-            LocalDate.parse(test, DateTimeFormatter.ofPattern("dd-MM-yyyy"));
-        } catch (DateTimeParseException e) {
-            return false;
-        }
-        return true;
-    }
     @Override
     public String toString() {
         return "Upcoming Appointment Date: " + (value == null ? "None" : value);
